Ivy - initial setup
In an effort to move to Ivy for dependency / library management. At
the moment it works and is setup.  However there still are questions
and issues that I have not surmounted.

1) How do I work offline?
2) How can I trim the download?
   -Right now it pulls down source jars and javadoc jars
   -The transient dependencies seem a bit much and not all needed.

I would also like to build my own production and development repos for
build artifacts that each project generates.  And have them be posted
to said repos.

As I said, *initial* setup of Ivy. :-)