Factored out some vars to make more universal
moved the "module" property out of init into the global scope and then
using it everywhere that module string was located.  Also, created an
author property and replaced that in the manifests for the various
jars that are created.  This is in effort to make the build script
more portable for new projects and authors.  There will still be some
manual tweaking to the script that has to be done, but this gets us
closer to minimizing those tweaks.

-GMB